separate api from http client
